requirements.txt
pandas
numpy
sentence-transformers
faiss-cpu
scikit-learn
scipy
transformers
datasets
tqdm
matplotlib
plotly
